rename user handler files to user lookup
Just file name changes & import changes for this commit
