+++ src/jgikbase/idmapping/cli.py
@@ -9,7 +9,7 @@ from jgikbase.idmapping.core.arg_check import not_none, no_Nones_in_iterable
 import traceback
 from pathlib import Path
 from jgikbase.idmapping.builder import IDMappingBuilder
-from jgikbase.idmapping.core.user_handler import LocalUserHandler
+from jgikbase.idmapping.core.user_lookup import LocalUserHandler
 from jgikbase.idmapping.core.user import Username
